improve handle of USE_PARALLEL
